Add webpack and use it
Anything in fe is parsed by webpack. The entry point is application.jsx so only
things mentioned in that file will be processed. The output is a bundle.js file
which is put in app/assets/javascripts to be used by Sprockets.

This means that we can now use ES6 imports and CommonJS modules easily, just
`npm install --save foo` and then `require('foo')` or `import Foo from 'foo'`
as you would if you were a Node developer.

Downside to this is that you need npm and webpack installed (install webpack
with `npm install -g webpack` as well as running `npm install` to get the
dependencies set up. Seems a small price to pay for being able to use all the
JS hotness.